Allow configuration settings using YAML
This patch allows using plain yaml as input for some of the X_CSI_*
configuration settings. So far these had to be valid JSON strings
and these were passed through to Ember itself. YAML maps will be
converted to JSON and forwarded to Ember as a string.

So instead of using this:

spec:
  config:
    envVars:
      X_CSI_BACKEND_CONFIG : '{"driver": "RBD", "rbd_user": "admin"}'

You can now simply use this:

spec:
  config:
    envVars:
      X_CSI_BACKEND_CONFIG:
        driver: RBD
        rbd_user: admin

This is much more user friendly and makes it easier to set correct
values. Using YAML it is also possible to use settings provided by
the Operatorhub forms.

This patch is backward-compatible. If settings are still provided
as a JSON string they will be checked and forwarded to Ember
itself.  If a plain string is provided it will be simply forwarded
to Ember itself.