Transform configuration options
Adding a function to translate configuration values according to the
suffix of the corresponding key. These operations are applied to the
JSON config both on new and existing deployments using the reconciler.

Currently supported transformation suffixes:
  __transform_empty_none
  __transform_csv
  __transform_csv_kvs

__transform_empty_none replaces an empty string with a null value.

__transform_csv splits a comma-separated string list into a JSON list.

__transform_csv_kvs splits a comma-separated list of key-value pairs
(separated with a ":") into a JSON dictionary.

The suffixes are removed from the key names after transforming.