Remove unused driver configuration settings
Each driver setting can now be prefixed like this:

  driver__DRIVER_NAME__

All settings will be removed from the config if the DRIVER_NAME does not
match the selected driver. If it matchtes, the prefix will be removed,
only keeping the setting name itself.

This reduces the active YAML configuration quite a bit, making it much
easier for operators to modify and maintain settings as needed.