Set X_CSI_BACKEND_CONFIG/name and X_CSI_EMBER_CONFIG/plugin_name
Adds a new function to update a JSON string and set a given key if empty
or unset. This function is used to set X_CSI_BACKEND_CONFIG/name and
X_CSI_EMBER_CONFIG/plugin_name to the name of the EmberCSI object if not
set.  Additionally plugin_name has to be unique, thus adding a random
suffix of 6 chars to the name itself.