Set defaults for X_CSI_PERSISTENCE_CONFIG
The namespace was only set if X_CSI_PERSISTENCE_CONFIG was not set at
all, but not if only the backend was set. This patch sets both the
storage backend and the namespace correctly if not defined.

Also fix a small bug alongside, where an invalid JSON string was
returned when an invalid JSON was parsed by interfaceToString.