UI: Add missing RBD parameter
The rbd_keyring_conf parameter has been marked as deprecated in Cinder
because it should not be used within OpenStack, as users there can see
the connection information and would then gain access to the whole
cluster.

But in Ember-CSI is completely different, because here normal users
should not have access to the Ember-CSI CRD objects where this
information is kept.

Since we are removing deprecated options by default this option was
removed from the form and examples, so in this patch we forcefully add
it in the generator.