Update YAML auto generation script
Our current olm-csv-gen.sh script has a mix of podman and docker
commands.  We should stick to one of them, so this patch changes the
`docker run` command for `podman run`.

In the same script container images are not referencing a specific
repository, so they may go to docker.io first, which we'll no longer be
using now that we have switched to quay.io.  This patch makes the images
repository explicit.